% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ids.R
\name{clean_ids}
\alias{clean_ids}
\title{Clean and Save IDS Surveillance Data}
\usage{
clean_ids(
  data_file = NULL,
  data_root = here::here("data", "ids"),
  maladie = "measles",
  year = as.numeric(format(Sys.Date(), "\%Y")),
  max_week = NULL,
  add_history = TRUE,
  save = TRUE,
  output_root = here::here("out")
)
}
\arguments{
\item{data_file}{\code{str} File to be loaded and cleaned.}

\item{data_root}{\code{str} Directory where data is stored. Default is 'data'.}

\item{maladie}{\code{str} Maladie to extract and clean.}

\item{year}{\code{int} Year to associate with the data in \code{data_file}. Default is the current year.}

\item{max_week}{\code{int} The maximum week expected in the dataset.}

\item{add_history}{\code{bool} If \code{TRUE}, historic IDS data (from 2019) will be added included in the
output. Default is \code{TRUE}.}

\item{save}{\code{bool} If \code{TRUE}, the output data will be saved to a file. Default is \code{TRUE}.}

\item{output_root}{\code{str} Path to be used when saving the output data (if \code{save} is \code{TRUE}). Default
is 'current_clean.csv'.}
}
\value{
df
}
\description{
Loads and cleans IDS data with the option to save the cleaned output to a file in
addition to loading it into the environment.
}
